refactor:이모지 컴포넌트 분리
