+++ src/components/EmojiAddition/EmojiAddition.module.css
@@ -1,0 +1,5 @@
+.emojiContainer {
+  display: flex;
+  flex-direction: row;
+  gap: 10px;
+}
