+++ src/pages/StudyDetail/StudyDetail.module.css
@@ -13,11 +13,7 @@
   align-items: center;
   justify-content: space-between;
 }
-.emojiContainer {
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-}
+
 .navContainer {
   display: flex;
   flex-direction: row;
